+++ Base/include/DLineArith_SSE.h
@@ -44,12 +44,12 @@
 template <>
 struct addLine<UINT8> : public binaryLineFunctionBase<UINT8>
 {
-    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
         for (int i=0;i<size;i++)
             lOut[i] = lIn1[i] > (UINT8)(numeric_limits<UINT8>::max()- lIn2[i]) ? numeric_limits<UINT8>::max() : lIn1[i] + lIn2[i];
     }
-    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
 	__m128i r0,r1;
 	__m128i *l1 = (__m128i*) lIn1;
@@ -58,7 +58,7 @@ struct addLine<UINT8> : public binaryLineFunctionBase<UINT8>
 	
 	unsigned long alignLen = size-size%SIMD_VEC_SIZE;
 	
-	for(UINT i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
+	for(size_t i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
 	{
 	    r0 = _mm_load_si128(l1);
 	    r1 = _mm_load_si128(l2);
@@ -73,12 +73,12 @@ struct addLine<UINT8> : public binaryLineFunctionBase<UINT8>
 template <>
 struct addNoSatLine<UINT8> : public binaryLineFunctionBase<UINT8>
 {
-    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
         for (int i=0;i<size;i++)
             lOut[i] = lIn1[i] + lIn2[i];
     }
-    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
 	__m128i r0,r1;
 	__m128i *l1 = (__m128i*) lIn1;
@@ -87,7 +87,7 @@ struct addNoSatLine<UINT8> : public binaryLineFunctionBase<UINT8>
 	
 	unsigned long alignLen = size-size%SIMD_VEC_SIZE;
 	
-	for(UINT i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
+	for(size_t i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
 	{
 	    r0 = _mm_load_si128(l1);
 	    r1 = _mm_load_si128(l2);
@@ -102,12 +102,12 @@ struct addNoSatLine<UINT8> : public binaryLineFunctionBase<UINT8>
 template <>
 struct subLine<UINT8> : public binaryLineFunctionBase<UINT8>
 {
-    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
         for (int i=0;i<size;i++)
             lOut[i] = lIn1[i] < (UINT8)(numeric_limits<UINT8>::max() + lIn2[i]) ? numeric_limits<UINT8>::min() : lIn1[i] - lIn2[i];
     }
-    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
 	__m128i r0,r1;
 	__m128i *l1 = (__m128i*) lIn1;
@@ -116,7 +116,7 @@ struct subLine<UINT8> : public binaryLineFunctionBase<UINT8>
 	
 	unsigned long alignLen = size-size%SIMD_VEC_SIZE;
 	
-	for(UINT i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
+	for(size_t i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
 	{
 	    r0 = _mm_load_si128(l1);
 	    r1 = _mm_load_si128(l2);
@@ -131,12 +131,12 @@ struct subLine<UINT8> : public binaryLineFunctionBase<UINT8>
 template <>
 struct subNoSatLine<UINT8> : public binaryLineFunctionBase<UINT8>
 {
-    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
         for (int i=0;i<size;i++)
             lOut[i] = lIn1[i] - lIn2[i];
     }
-    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
 	__m128i r0,r1;
 	__m128i *l1 = (__m128i*) lIn1;
@@ -145,7 +145,7 @@ struct subNoSatLine<UINT8> : public binaryLineFunctionBase<UINT8>
 	
 	unsigned long alignLen = size-size%SIMD_VEC_SIZE;
 	
-	for(UINT i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
+	for(size_t i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
 	{
 	    r0 = _mm_load_si128(l1);
 	    r1 = _mm_load_si128(l2);
@@ -160,12 +160,12 @@ struct subNoSatLine<UINT8> : public binaryLineFunctionBase<UINT8>
 template <>
 struct supLine<UINT8> : public binaryLineFunctionBase<UINT8>
 {
-    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
         for (int i=0;i<size;i++)
             lOut[i] = lIn1[i] > lIn2[i] ? lIn1[i] : lIn2[i];
     }
-    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
 	__m128i r0,r1;
 	__m128i *l1 = (__m128i*) lIn1;
@@ -174,7 +174,7 @@ struct supLine<UINT8> : public binaryLineFunctionBase<UINT8>
 	
 	unsigned long alignLen = size-size%SIMD_VEC_SIZE;
 	
-	for(UINT i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
+	for(size_t i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
 	{
 	    r0 = _mm_load_si128(l1);
 	    r1 = _mm_load_si128(l2);
@@ -189,12 +189,12 @@ struct supLine<UINT8> : public binaryLineFunctionBase<UINT8>
 template <>
 struct infLine<UINT8> : public binaryLineFunctionBase<UINT8>
 {
-    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
         for (int i=0;i<size;i++)
             lOut[i] = lIn1[i] < lIn2[i] ? lIn1[i] : lIn2[i];
     }
-    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, int size, UINT8 *lOut)
+    inline void _exec_aligned(UINT8 *lIn1, UINT8 *lIn2, size_t size, UINT8 *lOut)
     {
 	__m128i r0,r1;
 	__m128i *l1 = (__m128i*) lIn1;
@@ -203,7 +203,7 @@ struct infLine<UINT8> : public binaryLineFunctionBase<UINT8>
 	
 	unsigned long alignLen = size-size%SIMD_VEC_SIZE;
 	
-	for(UINT i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
+	for(size_t i=0 ; i<alignLen ; i+=16, l1++, l2++, l3++)
 	{
 	    r0 = _mm_load_si128(l1);
 	    r1 = _mm_load_si128(l2);
